June Code Push...
Changed the way account settings are saved.
Change the DI Container from Bucket to something simpler.
Fix up the FileCache and the ApcCache.
After getting burned by whitespace after ?> I have removed all the
unnecessary trailing ?> in the php files.
Rewrite the Model generator so that we have automated fields generated
as well as properly documented Retreive, RetreiveMultiple and create
fields.
Add the beginnings of tests for the KrisCronManager.